Moved CohortCatMongo to subdocuments package
